requirements.txt
abnf==2.2.0
accelerate==1.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
backoff==2.2.1
bitsandbytes==0.45.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
cint==1.0.0
click==8.3.1
cryptography==46.0.3
datasets==3.2.0
dill==0.3.8
diskcache==5.6.3
docstring_parser==0.17.0
einops==0.8.1
exceptiongroup==1.3.1
fickling==0.1.7
filelock==3.20.3
flash_attn==2.8.3
frozenlist==1.8.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.46
gql==4.0.0
graphql-core==3.2.7
graphviz==0.21
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
intervaltree==3.2.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaitaistruct==0.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.2.3
pdfminer.six==20260107
peft==0.14.0
pillow==11.3.0
platformdirs==4.5.1
polyfile-weave==0.5.8
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pyarrow==23.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.50.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
stdlib-list==0.11.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.50.0
triton==3.1.0
trl==0.12.0
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.5
tzdata==2025.3
urllib3==2.6.3
wandb==0.24.0
weave==0.52.25
xxhash==3.6.0
yarl==1.22.0
